ajouter retirer membre

+++ harmony-api/src/main/java/com/example/harmonyapi/jpa/postgresql/GroupUserController.java
@@ -14,6 +14,9 @@ public class GroupUserController {
     @Autowired
     GroupUserRepository groupUserRepository;
 
+    @Autowired
+    UserRepository userRepository;
+
     @GetMapping()
     public ResponseEntity<GroupUser> getGroupUserByPseudoAndIdGroup(@RequestParam(required = true) String pseudoUser, @RequestParam(required = true) int idGroup) {
         try {
@@ -27,6 +30,10 @@ public class GroupUserController {
     @PostMapping()
     public ResponseEntity<GroupUser> addGroupUser(@RequestBody GroupUser groupUser) {
         try {
+            if (userAlreadyInGroup(groupUser.getIdGroup(), groupUser.getPseudoUser())) {
+                return new ResponseEntity<>(null, HttpStatus.CONFLICT);
+            }
+
             GroupUser savedGroupUser = groupUserRepository.save(groupUser);
             return new ResponseEntity<>(savedGroupUser, HttpStatus.CREATED);
         } catch (Exception e) {
@@ -34,9 +41,23 @@ public class GroupUserController {
         }
     }
 
+    private boolean userAlreadyInGroup(long idGroup, String pseudoUser) {
+        return groupUserRepository.findByPseudoUserAndIdGroup(pseudoUser, idGroup) != null;
+    }
+
+    private boolean userDoesntExist(String pseudoUser) {
+        return userRepository.findByPseudo(pseudoUser).isEmpty();
+    }
+
     @DeleteMapping()
     public ResponseEntity<Expense> deleteGroupUser(@RequestParam(required = true) String pseudoUser, @RequestParam(required = true) int idGroup) {
         try {
+            if(userDoesntExist(pseudoUser)) {
+                return new ResponseEntity<>(null, HttpStatus.NOT_FOUND);
+            }
+            if(!userAlreadyInGroup(idGroup, pseudoUser)) {
+                return new ResponseEntity<>(null, HttpStatus.FORBIDDEN);
+            }
             GroupUser groupUser = groupUserRepository.findByPseudoUserAndIdGroup(pseudoUser, idGroup);
             groupUserRepository.delete(groupUser);
             return new ResponseEntity<>(null, HttpStatus.OK);
